Reorganize cb module
Renaming files according to issue #12.
