+++ cb/cb.py
@@ -1,189 +1,77 @@
-import os
 import math
-
-import magma as m
-import mantle as mantle
+from bit_vector import BitVector
 
 
-def run_cmd(cmd):
-        res = os.system(cmd)
-        assert(res == 0)
-
-
-def make_name(width, num_tracks, has_constant, default_value,
-              feedthrough_outputs):
-    return (f"connect_box_width_width_{width}"
-            f"_num_tracks_{num_tracks}"
-            f"_has_constant{has_constant}"
-            f"_default_value{default_value}"
-            f"_feedthrough_outputs_{feedthrough_outputs}")
-
-
-def generate_inputs(num_tracks, feedthrough_outputs, width):
-    """
-    Example feedthrough_outputs parameter:
-        "feedthrough_outputs": "1111101111",
-    """
-    IO = []
-    for i in range(0, num_tracks):
-        if (feedthrough_outputs[i] == '1'):
-            IO.append(f"in_{i}")
-            IO.append(m.In(m.Bits(width)))
-    return IO
-
-
-def generate_reset_value(constant_bit_count, default_value, reset_val,
-                         mux_sel_bit_count):
-    config_reg_reset_bit_vector = []
-
-    if (constant_bit_count > 0):
-        print('constant_bit_count =', constant_bit_count)
-
-        reset_bits = m.bitutils.int2seq(default_value,
-                                        constant_bit_count)
-        default_bits = m.bitutils.int2seq(reset_val, mux_sel_bit_count)
-
-        print('default val bits =', reset_bits)
-        print('reset val bits   =', default_bits)
-
-        # concat before assert
-        config_reg_reset_bit_vector += default_bits
-        config_reg_reset_bit_vector += reset_bits
-
-        config_reg_reset_bit_vector = \
-            m.bitutils.seq2int(config_reg_reset_bit_vector)
-        print('reset bit vec as int =', config_reg_reset_bit_vector)
-
-    else:
-        config_reg_reset_bit_vector = reset_val
-
-    return config_reg_reset_bit_vector
-
-
-def generate_output_mux(num_tracks, feedthrough_outputs, has_constant, width,
-                        mux_sel_bit_count, constant_bit_count, io,
-                        config_register):
-    pow_2_tracks = 2**m.bitutils.clog2(num_tracks)
-    print('# of tracks =', pow_2_tracks)
-    output_mux = mantle.Mux(height=pow_2_tracks, width=width)
-    m.wire(output_mux.S, config_register.O[:m.bitutils.clog2(pow_2_tracks)])
-
-    # This is only here because this is the way the switch box numbers
-    # things.
-    # We really should get rid of this feedthrough parameter
-    sel_out = 0
-    for i in range(0, pow_2_tracks):
-        # in_track = 'I' + str(i)
-        if (i < num_tracks):
-            if (feedthrough_outputs[i] == '1'):
-                m.wire(getattr(output_mux, 'I' + str(sel_out)),
-                       getattr(io, 'in_' + str(i)))
-                sel_out += 1
-
-    if (has_constant == 0):
-        while (sel_out < pow_2_tracks):
-            m.wire(getattr(output_mux, 'I' + str(sel_out)), m.uint(0, width))
-            sel_out += 1
-    else:
-        const_val = config_register.O[
-            mux_sel_bit_count:
-            mux_sel_bit_count + constant_bit_count
-        ]
-        while (sel_out < pow_2_tracks):
-            m.wire(getattr(output_mux, 'I' + str(sel_out)), const_val)
-            sel_out += 1
-    return output_mux
-
-
-@m.cache_definition
-def define_cb(width, num_tracks, has_constant, default_value,
-              feedthrough_outputs):
-    CONFIG_DATA_WIDTH = 32
+def gen_cb(width: int,
+           num_tracks: int,
+           feedthrough_outputs: str,
+           has_constant: bool,
+           default_value: int):
     CONFIG_ADDR_WIDTH = 32
+    CONFIG_DATA_WIDTH = 32
+    # The names of the inputs to the module are given by @feedthrough_outputs.
+    inputs = ["in_" + str(i)
+              for i, c in enumerate(feedthrough_outputs) if c == "1"]
+    # The number of total config bits needed is the number of bits needed to
+    # select between all the inputs (including the constant if it is included),
+    # as well as a constant value if @has_constant is True. Consequently, the
+    # number of config registers needed is the total number of config bits
+    # needed, divided by the size of a single config data element (e.g. 32
+    # bits).
+    mux_height = len(inputs)
+    if has_constant:
+        mux_height += 1
+    mux_sel_bits = math.ceil(math.log(mux_height, 2))
+    config_bits_needed = mux_sel_bits
+    if has_constant:
+        config_bits_needed += width
+    num_config_regs = math.ceil(config_bits_needed / CONFIG_DATA_WIDTH)
 
-    constant_bit_count = has_constant * width
-    feedthrough_count = num_tracks
-    for i in range(0, len(feedthrough_outputs)):
-        feedthrough_count -= feedthrough_outputs[i] == '1'
+    class _Cb:
+        def __init__(self):
+            self.__config = [BitVector(0, CONFIG_DATA_WIDTH)
+                             for _ in range(num_config_regs)]
 
-    reset_val = num_tracks - feedthrough_count + has_constant - 1
+        def configure(self, addr: BitVector, data: BitVector):
+            assert addr.num_bits == CONFIG_ADDR_WIDTH
+            assert data.num_bits == CONFIG_ADDR_WIDTH
+            addr_high_bits = addr[24:32]
+            config_reg_select = addr_high_bits.as_int()
+            if config_reg_select in range(num_config_regs):
+                self.__config[config_reg_select] = data
 
-    mux_sel_bit_count = m.bitutils.clog2(
-        num_tracks - feedthrough_count + has_constant)
+        # Function to slice the global config bit space. Returns bits in the
+        # range [lo, hi).
+        def __get_config_bits(self, lo: int, hi: int):
+            assert hi > lo
+            assert lo >= 0
+            assert hi <= (len(self.__config) * CONFIG_DATA_WIDTH)
+            start = math.floor(lo / 32)
+            end = math.floor((hi - 1) / 32)
+            lo_int = lo % CONFIG_DATA_WIDTH
+            hi_int = hi % CONFIG_DATA_WIDTH
+            if start == end:
+                return self.__config[start][lo_int:hi_int]
+            ret = self.__config[start][lo_int:CONFIG_DATA_WIDTH]
+            for i in range(start + 1, end):
+                ret = BitVector.concat(ret, self.__config[i])
+            ret = BitVector.concat(ret, self.__config[i][0:hi_int])
+            assert ret.num_bits == (hi - lo)
+            return ret
 
-    config_bit_count = mux_sel_bit_count + constant_bit_count
+        def __call__(self, *args):
+            assert len(args) == len(inputs)
+            select = self.__get_config_bits(0, mux_sel_bits)
+            select_as_int = select.as_int()
+            if select_as_int in range(len(inputs)):
+                return args[select_as_int]
+            # TODO(raj): Handle the case where we select the constant value or
+            # resort to default.
+            raise Exception()
 
-    config_reg_width = int(math.ceil(config_bit_count / 32.0)*32)
+        # Debug method to read config data.
+        @property
+        def config(self):
+            return self.__config
 
-    class ConnectBox(m.Circuit):
-        name = make_name(width, num_tracks, has_constant, default_value,
-                         feedthrough_outputs)
-
-        # TODO: We chose to use explicit clock interface here so the names
-        # match the genesis verilog for regression testing, this should really
-        # use m.ClockInterface
-        IO = ["clk", m.In(m.Clock),
-              "reset", m.In(m.Reset)]
-
-        IO += generate_inputs(num_tracks, feedthrough_outputs, width)
-
-        IO += [
-            "out", m.Out(m.Bits(width)),
-            "config_addr", m.In(m.Bits(CONFIG_ADDR_WIDTH)),
-            "config_data", m.In(m.Bits(CONFIG_DATA_WIDTH)),
-            "config_en", m.In(m.Bit),
-            "read_data", m.Out(m.Bits(CONFIG_DATA_WIDTH))
-        ]
-
-        @classmethod
-        def definition(io):
-
-            config_reg_reset_bit_vector = \
-                generate_reset_value(constant_bit_count, default_value,
-                                     reset_val, mux_sel_bit_count)
-
-            config_cb = mantle.Register(config_reg_width,
-                                        init=config_reg_reset_bit_vector,
-                                        has_ce=True,
-                                        has_reset=True)
-
-            config_addr_zero = mantle.eq(m.uint(0, 8), io.config_addr[24:32])
-
-            config_cb(io.config_data, reset=io.reset,
-                      CE=io.config_en & config_addr_zero)
-
-            # if the top 8 bits of config_addr are 0, then read_data is equal
-            # to the value of the config register, otherwise it is 0
-            m.wire(io.read_data,
-                   mantle.mux([m.uint(0, 32), config_cb.O],
-                              config_addr_zero))
-
-            output_mux = generate_output_mux(num_tracks, feedthrough_outputs,
-                                             has_constant, width,
-                                             mux_sel_bit_count,
-                                             constant_bit_count, io, config_cb)
-
-            # NOTE: This is a dummy! fix it later!
-            m.wire(output_mux.O, io.out)
-            return
-
-    return ConnectBox
-
-
-def generate_genesis_cb(p_width,
-                        num_tracks,
-                        feedthrough_outputs,
-                        has_constant,
-                        default_value):
-        run_cmd('Genesis2.pl -parse -generate ' +
-                '-top cb -input ./tests/cb.vp ' +
-                '-parameter cb.width=' +
-                str(p_width) +
-                ' -parameter cb.num_tracks=' +
-                str(num_tracks) +
-                ' -parameter cb.has_constant=' +
-                str(has_constant) +
-                ' -parameter cb.default_value=' +
-                str(default_value) +
-                ' -parameter cb.feedthrough_outputs=' +
-                feedthrough_outputs)
+    return _Cb
