+++ cb/cb_genesis2.py
@@ -1,9 +1,16 @@
-"""
-This program generates the verilog for the connect box and parses it into a
-Magma circuit. The circuit declaration is printed at the end of the program.
-"""
 import argparse
-from cb.cb_wrapper import define_cb_wrapper
+from common.genesis_wrapper import define_genesis_generator
+
+
+define_cb_wrapper = define_genesis_generator(
+    top_name="cb",
+    input_files=["cb.vp"],
+    width=16,
+    num_tracks=10,
+    feedthrough_outputs="1111101111",
+    has_constant=1,
+    default_value=0
+)
 
 
 def create_parser():
@@ -18,7 +25,6 @@ def create_parser():
 
 
 def main(args):
-
     # Check some of the inputs.
     for c in args.feedthrough_outputs:
         assert c == "1" or c == "0"
@@ -32,6 +38,10 @@ def main(args):
     print(cb)
 
 
+"""
+This program generates the verilog for the connect box and parses it into a
+Magma circuit. The circuit declaration is printed at the end of the program.
+"""
 if __name__ == "__main__":
     # These functions are unit tested directly, so no need to cover them
     parser = create_parser()  # pragma: no cover
