+++ test_cb/test_regression.py
@@ -4,9 +4,9 @@ import os
 import glob
 from bit_vector import BitVector
 
-from cb.cb import define_cb
-from cb.cb_functional_model import gen_cb
-from cb.cb_wrapper import define_cb_wrapper
+from cb.cb_magma import define_cb
+from cb.cb import gen_cb
+from cb.cb_genesis2 import define_cb_wrapper
 
 import magma as m
 import fault
@@ -46,7 +46,7 @@ def test_regression(default_value, num_tracks, has_constant):
     magma_cb = define_cb(**params)
     compile_to_verilog(magma_cb, magma_cb.name, "test_cb/build/")
 
-    genesis_cb = define_cb_wrapper(**params, input_files=["test_cb/cb.vp"])
+    genesis_cb = define_cb_wrapper(**params, input_files=["cb/cb.vp"])
     genesis_verilog = "genesis_verif/cb.v"
     shutil.copy(genesis_verilog, "test_cb/build")
